requirements.txt
asgiref==3.5.0
autopep8==1.6.0
beautifulsoup4==4.11.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap4==22.1
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-registration==3.2
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
numpy==1.22.3
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
soupsieve==2.3.2
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.0.0
zipp==3.8.0
